Update README.md
LINKED LIST IMPLEMENTATION IN A GUI AS A JAVAFX APPLICATION.

Linked List is a singly linear linked list implementation of Java’s List. It is part of Java’s collections framework. The framework will animate your code to give you feedback on the correctness of your code. Displays state of the linked list.